+++ lib/AppInfo/Application.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace OCA\BerthaWebhook\AppInfo;
+namespace OCA\BerthaKi\AppInfo;
 
-use OCA\BerthaWebhook\Listener\ChatMessageListener;
-use OCA\BerthaWebhook\Listener\FileShareListener;
-use OCA\BerthaWebhook\Listener\ReactionListener;
+use OCA\BerthaKi\Listener\ChatMessageListener;
+use OCA\BerthaKi\Listener\FileShareListener;
+use OCA\BerthaKi\Listener\ReactionListener;
 use OCA\Talk\Events\ChatMessageSentEvent;
 use OCA\Talk\Events\ReactionAddedEvent;
 use OCP\AppFramework\App;
@@ -17,7 +17,7 @@ use OCP\Share\Events\ShareCreatedEvent;
 
 class Application extends App implements IBootstrap {
 
-	public const APP_ID = 'bertha_webhook';
+	public const APP_ID = 'bertha_ki';
 
 	public function __construct() {
 		parent::__construct(self::APP_ID);